COMP: many warnings removed. STYLE: mainly indentation.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=8192
git-svn-id: http://svn.slicer.org/Slicer4/trunk@8192 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/Qdec/QdecProject.cpp
@@ -933,7 +933,7 @@ int QdecProject::GenerateMappedLabelForAllSubjects
       status += "'...";
       iProgressUpdateGUI->UpdateProgressMessage( status.c_str() );
       iProgressUpdateGUI->UpdateProgressPercent
-        ( (float)nStep++ * stepIncrement );
+        ( (double)nStep++ * stepIncrement );
     }
     char* sCommand = strdup( ssCommand.str().c_str() );
     printf( "\n----------------------------------------------------------\n" );
